+++ src/ui/mod.rs
@@ -9,13 +9,12 @@
 #![cfg_attr(feature = "safe_ui", forbid(unsafe_code))]
 // docs
 crate::CONST! { pub(crate) _DOC_UI_MODULES =
-    crate::_doc!(modules: crate; ui: event, frame, layout); // intent, layout, view
+    crate::_doc!(modules: crate; ui: event, frame, layout, route); // intent, view, widget
 }
 
-// mod error; // RETHINK
+// mod error;
 mod intent; // WIP
 // #[cfg(feature = "layout")]
-mod route; // WIP
 mod sem; // WIP
 mod view; // WIP
 mod widget; // WIP
@@ -23,14 +22,14 @@ mod widget; // WIP
 #[cfg(feature = "event")]
 pub mod event; // Event[Button[State]|Key[State]|Kind|Mouse|Pointer[Type]|TimeStamp|Wheel], Key*
 pub mod frame; // UiFrame, UiId, UiKey, UiPhase, UiScope
-pub mod layout; // Layout1d, LayoutReceipt, Lunit, …
+pub mod layout; // Layout1d, LayoutReceipt, Lunit, Layout<Pos*|Ext*|Rec|Region|Stride*|>
+pub mod route; // HitRoute, Route<Active|Capture|Focus|Hot>
 
 crate::structural_mods! { // _mods, _pub_mods, _crate_internals
     _mods {
         pub use super::{
             // error::*,
             intent::*,
-            route::*,
             sem::*,
             view::*,
             widget::*,
@@ -40,6 +39,7 @@ crate::structural_mods! { // _mods, _pub_mods, _crate_internals
         pub use super::{
             frame::*,
             layout::*,
+            route::*,
         };
         #[cfg(feature = "event")]
         pub use super::event::_all::*;